requirements.txt
requests
beautifulsoup4
pandas
pymysql
sqlalchemy
openpyxl
lxml
psycopg2-binary
pororo-ocr
